Renamed adapter to adaptor to ensure consistency within code.
