+++ mhs/mhs/message_configuration/lookup/dictionary_cache.py
@@ -1,7 +1,7 @@
 import time
 from typing import Dict, Optional
 
-import mhs.message_configuration.lookup.cache_adapter as cache_adapter
+import mhs.message_configuration.lookup.cache_adaptor as cache_adaptor
 
 FIFTEEN_MINUTES_IN_SECONDS = 900
 
@@ -10,7 +10,7 @@ def _generate_key(ods_code: str, interaction_id: str) -> str:
     return ods_code + '-' + interaction_id
 
 
-class DictionaryCache(cache_adapter.CacheAdapter):
+class DictionaryCache(cache_adaptor.CacheAdaptor):
 
     def __init__(self, expiry_time: float = FIFTEEN_MINUTES_IN_SECONDS):
         if expiry_time < 0:
